--- srcs/socket/SocketMap.hpp
@@ -1,31 +1,0 @@
-#ifndef SRCS_SOCKET_SOCKETMAP_HPP
-#define SRCS_SOCKET_SOCKETMAP_HPP
-
-#include <poll.h>
-
-#include "config/Config.hpp"
-#include "socket/SocketBase.hpp"
-#include "socket/SocketMapOp.hpp"
-
-class SocketMap {
-private:
-  serverList                  _server_list_;
-  std::map<int, SocketBase *> _socket_map_;
-
-private:
-  SocketMap();
-  SocketMap(const SocketMap &other);
-  SocketMap                  &operator=(const SocketMap &other);
-  void                        _read_config(const char *config_file_path);
-  std::map<int, SocketBase *> _generate();
-
-public:
-  SocketMap(const char *config_file_path);
-  ~SocketMap(){};
-  std::vector<struct pollfd> pollfds();
-  SocketMapOp handle_socket_event(int socket_fd, short int revents);
-  void        do_map_operation(const SocketMapOp &socket_map_op);
-  void        close_timedout_socket();
-};
-
-#endif /* SRCS_SOCKET_SOCKETMAP_HPP */
