+++ srcs/webserv.cpp
@@ -4,7 +4,6 @@
 
 #include "config/Config.hpp"
 #include "event/Server.hpp"
-#include "socket/SocketMap.hpp"
 
 #define DEFAULT_CONFIG_FILE_PATH "conf/webserv.conf"
